feat: make invocation of tools more declaritive
